bot/requirements.txt
aiogram==3.15.0
aiogram-dialog==2.2.0
aiohttp==3.11.10
python-dotenv==1.0.1